requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
pillow==11.2.1
psycopg2-binary==2.9.9
python-decouple
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
google-generativeai>=0.3.0
drf-yasg
django-widget-tweaks
dj-database-url==2.1.0
whitenoise==6.0.0
